requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
numpy==1.23.3
requests==2.24.0
gunicorn==20.1.0
Pillow==8.3.2
tensorflow-cpu==2.10.0
opencv-contrib-python-headless==4.6.0.66
pandas
scikit-learn
